pyproject.toml
[tool.poetry]
name = "sreality_scraper"
version = "0.1.0"
description = "Sreality scraper for scraping flats to sell from sreality.cz"
authors = ["Darja Timaseva <thedaryatim@gmail.com>"]

[[tool.poetry.packages]]
include = "app"
from = "."

[[tool.poetry.packages]]
include = "database"
from = "."

[[tool.poetry.packages]]
include = "http_server"
from = "."

[[tool.poetry.packages]]
include = "scraper"
from = "."

[[tool.poetry.packages]]
include = "tests"
from = "."


[tool.poetry.dependencies]
python = "^3.8"
pre-commit = "^3.5.0"
scrapy = "^2.11.0"
pydantic = "^2.5.2"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
redis = "^5.0.1"
sqlalchemy = "^2.0.23"
pytest = "^7.4.3"
pytest-postgresql = "^5.0.0"
sqlalchemy-utils = "^0.41.1"
pytest-alembic = "^0.10.7"
beautifulsoup4 = "^4.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
